+++ CHANGELOG.md
@@ -3,7 +3,8 @@
 ## Version 0.1.4
 
 - Added command `odoo-helper odoo`
-- Added command `odoo-helper odoo recompute`, that allows to recompute stored fields
+- Added command `odoo-helper odoo recompute`, that allows to recompute stored fields.
+  Also this command allow to recompute parents (left/right)
 - Command `odoo-helper db exists` now have it's own help message
 - Command `odoo-helper db exists` added option `-q` to disable output
 - Added command `odoo-helper postgres speedify`